+++ khoros/errors/exceptions.py
@@ -6,7 +6,7 @@
 :Example:           ``raise khoros.errors.exceptions.BadCredentialsError``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     01 Mar 2020
+:Modified Date:     21 Mar 2020
 """
 
 #################
@@ -23,6 +23,19 @@ class KhorosError(Exception):
 ############################
 # Authentication Exceptions
 ############################
+
+
+# Define exception for missing authentication data
+class InvalidCallbackURLError(KhorosError):
+    """This exception is used when an invalid Callback URL for OAuth 2.0 was not provided."""
+    def __init__(self, *args, **kwargs):
+        default_msg = "The OAuth 2.0 callback URL that was provided is invalid. The entire URL must be provided."
+        if not (args or kwargs):
+            args = (default_msg,)
+        elif 'val' in kwargs:
+            custom_msg = f"{default_msg.split('URL ')[0]}'{kwargs['val']}'{default_msg.split('URL')[1]}"
+            args = (custom_msg,)
+        super().__init__(*args)
 
 
 # Define exception for missing authentication data
@@ -58,6 +71,97 @@ class CurrentlyUnsupportedError(KhorosError):
         if not (args or kwargs):
             args = (default_msg,)
         # TODO: Add alternate message if variable is passed to the exception class
+        super().__init__(*args)
+
+
+#########################
+# Generic API Exceptions
+#########################
+
+
+class APIConnectionError(KhorosError):
+    """This exception is used when the API query could not be completed due to connection aborts and/or timeouts."""
+    def __init__(self, *args, **kwargs):
+        default_msg = "The API query could not be completed due to connection aborts and/or timeouts."
+        if not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class GETRequestError(KhorosError):
+    """This exception is used for generic GET request errors when there isn't a more specific exception."""
+    def __init__(self, *args, **kwargs):
+        default_msg = "The GET request did not return a successful response."
+        if not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class InvalidEndpointError(KhorosError):
+    """This exception is used when an invalid API endpoint / service is provided."""
+    def __init__(self, *args, **kwargs):
+        default_msg = "The supplied endpoint for the API is not recognized. (Examples of valid " + \
+                      "lookup types include 'people' and 'contents')"
+        if not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class InvalidLookupTypeError(KhorosError):
+    """This exception is used when an invalid API lookup type is provided."""
+
+    def __init__(self, *args, **kwargs):
+        default_msg = "The supplied lookup type for the API is not recognized. (Examples of valid " + \
+                      "lookup types include 'id' and 'email')"
+        if not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class InvalidRequestTypeError(KhorosError):
+    """This exception is used when an invalid API request type is provided."""
+
+    def __init__(self, *args, **kwargs):
+        default_msg = "The supplied request type for the API is not recognized. (Examples of valid " + \
+                      "request types include 'POST' and 'PUT')"
+        if not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class LookupMismatchError(KhorosError):
+    """This exception is used when an a lookup value doesn't match the supplied lookup type."""
+    def __init__(self, *args, **kwargs):
+        default_msg = "The supplied lookup type for the API does not match the value that was provided."
+        if not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class NotFoundResponseError(KhorosError):
+    """This exception is used when an API query returns a 404 response and there isn't a more specific class."""
+    def __init__(self, *args, **kwargs):
+        default_msg = "The API query returned a 404 response."
+        if not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class POSTRequestError(KhorosError):
+    """This exception is used for generic POST request errors when there isn't a more specific exception."""
+    def __init__(self, *args, **kwargs):
+        default_msg = "The POST request did not return a successful response."
+        if not (args or kwargs):
+            args = (default_msg,)
+        super().__init__(*args)
+
+
+class PUTRequestError(KhorosError):
+    """This exception is used for generic PUT request errors when there isn't a more specific exception."""
+    def __init__(self, *args, **kwargs):
+        default_msg = "The PUT request did not return a successful response."
+        if not (args or kwargs):
+            args = (default_msg,)
         super().__init__(*args)